requirements.txt
pyfiglet
pandas_datareader
datetime
numpy
matplotlib.pyplot
scipy.signal